add special init
